pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-security-gate"
version = "0.1.0"
description = "Zero-trust backdoor scanning, counterfactual testing, and detox dataset generation for object detectors."
requires-python = ">=3.9"
dependencies = [
  "ultralytics>=8.2.0",
  "opencv-python>=4.8.0",
  "Pillow>=10.0.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "PyYAML>=6.0.0",
  "tqdm>=4.66.0",
  "matplotlib>=3.7.0",
  "torch>=2.1.0"
]

[tool.setuptools]
packages = [
  "model_security_gate",
  "model_security_gate.adapters",
  "model_security_gate.cf",
  "model_security_gate.scan",
  "model_security_gate.detox",
  "model_security_gate.attack_zoo",
  "model_security_gate.guard",
  "model_security_gate.intake",
  "model_security_gate.utils",
  "model_security_gate.verify",
  "model_security_gate.report",
  "model_security_gate.t0"
]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["win-64", "linux-64", "osx-64", "osx-arm64"]

[tool.pixi.dependencies]
python = ">=3.9,<3.13"
pip = ">=24.0"
pytest = ">=8.0"

[tool.pixi.pypi-dependencies]
model-security-gate = { path = ".", editable = true }

[tool.pixi.tasks]
security-gate = "python scripts/security_gate.py"
strong-detox-yolo = "python scripts/strong_detox_yolo.py"
detox-build-dataset = "python scripts/detox_build_dataset.py"
detox-train-yolo = "python scripts/detox_train_yolo.py"
prune-channels = "python scripts/prune_channels.py"
score-detox-channels = "python scripts/score_detox_channels.py"
progressive-prune-yolo = "python scripts/progressive_prune_yolo.py"
train-feature-detox = "python scripts/train_feature_detox.py"
score-strong-channels = "python scripts/score_strong_channels.py"
runtime-guard = "python scripts/runtime_guard.py"
eval-yolo-metrics = "python scripts/eval_yolo_metrics.py"
acceptance-gate = "python scripts/acceptance_gate.py"
generate-report = "python scripts/generate_report.py"
asr-aware-detox-yolo = "python scripts/asr_aware_detox_yolo.py"
run-asr-regression = "python scripts/run_asr_regression.py"
asr-closed-loop-detox-yolo = "python scripts/asr_closed_loop_detox_yolo.py"
run-external-hard-suite = "python scripts/run_external_hard_suite.py"
hybrid-purify-detox-yolo = "python scripts/hybrid_purify_detox_yolo.py"
rnp-prune-yolo = "python scripts/rnp_prune_yolo.py"
pareto-merge-yolo = "python scripts/pareto_merge_yolo.py"
targeted-repair-yolo = "python scripts/targeted_repair_yolo.py"
oda-postnms-repair-yolo = "python scripts/oda_postnms_repair_yolo.py"
diagnose-oda-candidates = "python scripts/diagnose_oda_candidates.py"
oda-score-calibration-repair-yolo = "python scripts/oda_score_calibration_repair_yolo.py"
formal-intake = "python scripts/intake_model_card.py"
green-acceptance-gate = "python scripts/green_acceptance_gate.py"
joint-no-worse-scorecard = "python scripts/joint_no_worse_scorecard.py"
frontier-t1-autodetox = "python scripts/frontier_t1_autodetox_yolo.py"
semantic-surgical-repair = "python scripts/semantic_surgical_repair_yolo.py"
ci-help-smoke-all = "python scripts/ci_help_smoke_all.py"
check-heldout-leakage = "python scripts/check_heldout_leakage.py"

t0-benchmark-audit = "python scripts/t0_benchmark_audit.py"
t0-evidence-pack = "python scripts/t0_evidence_pack.py"
t0-frontier-plan = "python scripts/t0_frontier_plan.py"
t0-evidence-pipeline = "python scripts/t0_evidence_pipeline.py"
t0-poison-matrix-evidence = "python scripts/t0_poison_matrix_evidence.py"
t0-poison-matrix-completion-plan = "python scripts/t0_poison_matrix_completion_plan.py"
t0-poison-matrix-aggregate = "python scripts/t0_poison_matrix_aggregate.py"
t0-defense-leaderboard = "python scripts/t0_defense_leaderboard.py"
t0-defense-certificate = "python scripts/t0_defense_certificate.py"
t0-detox-ablation-plan = "python scripts/t0_detox_ablation_plan.py"
t0-leakage-audit = "python scripts/t0_leakage_audit.py"
t0-abs-scan = "python scripts/t0_abs_scan.py"
build-t0-attack-zoo-yolo = "python scripts/build_t0_attack_zoo_yolo.py"
build-t0-poison-train-dataset = "python scripts/build_t0_poison_train_dataset.py"
train-t0-poison-models-yolo = "python scripts/train_t0_poison_models_yolo.py"
plan-t0-poison-model-matrix = "python scripts/plan_t0_poison_model_matrix.py"
run-t0-multi-attack-detox-yolo = "python scripts/run_t0_multi_attack_detox_yolo.py"
t0-constrained-candidate-select = "python scripts/t0_constrained_candidate_select.py"
pytest = "python -m pytest -q"
ci-smoke = "python -m compileall -q model_security_gate scripts tests && python -m pytest -q"

[tool.pixi.system-requirements]
cuda = "13.1"

[tool.pixi.target.win-64.dependencies]
pytorch-gpu = "2.10.0.*"
torchvision = "0.25.0.*"
